server/requirements.txt
numpy>=1.19.0
Pillow>=8.0.0
scikit-learn>=0.24.0
